fix(dashboard): all comments
