+++ packages/dashboard/components/consent/hooks/useAuthInfo.ts
@@ -1,6 +1,7 @@
 import { useEffect, useState, useCallback } from 'react';
 import { IRelayPKP, SessionSigs } from '@lit-protocol/types';
-import validateSession from '../utils/validateSession';
+import validateSession from '../utils/getValidSessionSigs';
+import { disconnectWeb3 } from '@lit-protocol/auth-browser';
 // Define interfaces for the authentication info
 export interface AuthInfo {
   type: string;
@@ -16,6 +17,12 @@ export interface UseReadAuthInfo {
   sessionSigs: SessionSigs | null;
   isProcessing: boolean;
   error: string | null;
+}
+
+export interface UseSetAuthInfo {
+  setAuthInfo: (newAuthInfo: AuthInfo) => boolean;
+  updateAuthInfo: (updates: Partial<AuthInfo>) => boolean;
+  error: Error | null;
 }
 
 const AUTH_INFO_KEY = 'lit-auth-info';
@@ -61,15 +68,19 @@ export const useReadAuthInfo = (): UseReadAuthInfo => {
  * Hook to provide functions for setting and updating authentication info
  * @returns Functions to set and update auth info
  */
-export const useSetAuthInfo = () => {
+export const useSetAuthInfo = (): UseSetAuthInfo => {
+  const [error, setError] = useState<Error | null>(null);
+
   // Set new auth info (replaces existing)
   const setAuthInfo = useCallback((newAuthInfo: AuthInfo) => {
     try {
       localStorage.setItem(AUTH_INFO_KEY, JSON.stringify(newAuthInfo));
       return true;
-    } catch (error) {
+    } catch (err) {
+      const error = err as Error;
       console.error('Error storing auth info:', error);
-      return false;
+      setError(error);
+      throw error;
     }
   }, []);
 
@@ -83,25 +94,28 @@ export const useSetAuthInfo = () => {
         const currentAuthInfo = JSON.parse(storedAuthInfo) as AuthInfo;
         updatedAuthInfo = { ...currentAuthInfo, ...updates };
       } else {
-        // If creating new auth info without required fields, set defaults instead of throwing an error
         updatedAuthInfo = {
-          type: updates.type || (updates.agentPKP ? 'webauthn' : 'unknown'), // Default to webauthn if we have PKP info
-          authenticatedAt: updates.authenticatedAt || new Date().toISOString(),
+          type: updates.type,
+          authenticatedAt: updates.authenticatedAt,
           ...updates
         } as AuthInfo;
       }
       
       localStorage.setItem(AUTH_INFO_KEY, JSON.stringify(updatedAuthInfo));
       return true;
-    } catch (error) {
+    } catch (err) {
+      const error = err as Error;
       console.error('Error updating auth info:', error);
-      return false;
+      setError(error);
+      localStorage.removeItem(AUTH_INFO_KEY);
+      throw error;
     }
   }, []);
 
   return {
     setAuthInfo,
-    updateAuthInfo
+    updateAuthInfo,
+    error
   };
 };
 
@@ -110,18 +124,26 @@ export const useSetAuthInfo = () => {
  * @returns Function to clear auth info
  */
 export const useClearAuthInfo = () => {
+  const [error, setError] = useState<Error | null>(null);
+
   // Clear auth info from localStorage
-  const clearAuthInfo = useCallback(() => {
+  const clearAuthInfo = useCallback(async () => {
     try {
       localStorage.removeItem(AUTH_INFO_KEY);
+      await disconnectWeb3();
       return true;
-    } catch (error) {
+    } catch (err) {
+      const error = err as Error;
       console.error('Error clearing auth info:', error);
-      return false;
+      setError(error);
+      throw error;
     }
   }, []);
 
-  return clearAuthInfo;
+  return {
+    clearAuthInfo,
+    error
+  };
 };
 
 // For backward compatibility
